Fix a lost-message problem for delivered messages
When a message is passed to BusyBee using the "deliver" call, it is placed on a
queue to be returned via a subsequent "recv" call.  If a pause/unpause combo
comes in between these two calls, it was possible for the queue to become
corrupted.  All messages on the queue would be delivered, but all messages
placed on the queue after the corruption would be lost.  Eventually, recv would
drain the queue and fix the corruption, leading subsequently enqueued messages
to not appear lost.